update_3rdparty_deps.py: Consistent line breaks
The header lines in the diff do have terminating newlines, the actual
diff lines don't, if the input doesn't have it. The result was double
linebreaks (i.e. empty lines) in the headers of the diff output.

Fix by calling `splitlines()` with `keepends=True`, and eliding the
newline when joining.